+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 ### Added
 
 - Add .gitattributes file
+- Add TreeTools
+- Add deepCopy
 
 ### Changed
 
 - Publish to pub.dev
+- Basic implementation
+- Auto rename nodes on same keys
 
 ## [1.0.0] - 2025-07-12